+++ src/help.ts
@@ -3,9 +3,7 @@ import { BBHelpHelpWidget } from './help-widget';
 import { BBHelpHelpWidgetRenderer } from './help-widget-renderer';
 import { BBHelpStyleUtility } from './help-widget-style-utility';
 import { BBHelp } from './models/bbhelp';
-import { BBHelpAnalyticsService } from './service/analytics.service';
 import { BBHelpCommunicationService } from './service/communication.service';
-import { MixpanelKeys } from './service/mixpanel-keys';
 
 declare const BBHELP: any;
 
@@ -14,12 +12,9 @@ export abstract class BBHelpClient {
   public static initWidget(): BBHelp {
     const styleUtility = new BBHelpStyleUtility();
     const widgetRenderer = new BBHelpHelpWidgetRenderer();
-    const mixpanelKeys = new MixpanelKeys();
-    const analyticsService = new BBHelpAnalyticsService(mixpanelKeys);
     const communicationService = new BBHelpCommunicationService();
     const helpWidget = new BBHelpHelpWidget(
       widgetRenderer,
-      analyticsService,
       communicationService,
       styleUtility
     );